Rename pycookiecheat.py to chrome.py
And rename the test file to match.
