+++ tests/test_chrome.py
@@ -13,7 +13,7 @@ import pytest
 from playwright.sync_api import sync_playwright
 
 from pycookiecheat import chrome_cookies
-from pycookiecheat.pycookiecheat import get_linux_config, get_osx_config
+from pycookiecheat.chrome import get_linux_config, get_osx_config
 
 BROWSER = os.environ.get("TEST_BROWSER_NAME", "Chromium")
